requirements.txt
einops
numpy
Pillow
PyYAML
PyYAML
torch
torchvision
ipykernel
lpips
pyiqa
scikit-image